+++ CHANGELOG.md
@@ -13,7 +13,7 @@ For questions and support requests, please use
 
 ### New Features
 
-* Add support for BlueSky
+* Add support for BlueSky with hashtags and profile mentions
 * Add support for Devoxx events (including Voxxed Days)
 * Configure UI language via environment variable
 * Dynamically generate social headline based on enabled plugins